+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Rename project to `rj`
 - New syntax for navigating objects: `.data.id` instead of `["data"]["id"]`
+- Add `-r` option that's the same as `-o plain`
 
 ## v0.2.0